tactics-generator/requirements.txt
CairoSVG==2.7.1
chess==1.11.2
intervaltree==3.1.0
matplotlib==3.10.1
numpy==2.2.4
Pillow==11.1.0
Requests==2.32.3
tqdm==4.67.1
zstandard==0.23.0
